Addition of a check to make sure the user exists before trying to delete

+++ FlashFeetAPI/src/main/java/org/example/api/JobRoleService.java
@@ -33,9 +33,14 @@ public class JobRoleService {
     public int deleteJobRole(String jobRole) throws JobRoleDoesNotExistException, FailedToDeleteJobRoleException {
         try {
             JobRole deleteJobRole = jobRoleDao.getJobRole(jobRole);
+            if (deleteJobRole != null) {
+                jobRoleDao.deleteJobRole(jobRole);
 
-            jobRoleDao.deleteJobRole(jobRole);
-            return 1;
+                return 1;
+            }
+            else {
+                return -1;
+            }
         } catch (SQLException e) {
             System.err.println(e.getMessage());
             return -1;
